Refactored pokemon detail page to use a custom hook for data fetching
